refactor: Prodcut 컴포넌트 개선 - 재사용 가능한 컴포넌트들 분리 (카드, 썸네일 등) - 재사용 가능한 함수 util로 분리
