+++ src/components/Product/ProductEmpty/styles.module.scss
@@ -1,0 +1,7 @@
+.empty {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 20rem;
+  color: var(--color-secondary-400);
+}
