+++ src/components/Product/ProductThumbnail/styles.module.scss
@@ -1,0 +1,35 @@
+.cover {
+  overflow: hidden;
+  position: relative;
+  display: block;
+  margin-bottom: 1.6rem;
+  border-radius: 1.6rem;
+  background: var(--color-secondary-100);
+  transition: all 0.2s;
+
+  &:before {
+    content: "";
+    display: block;
+    width: 100%;
+    height: 0;
+    padding-bottom: 100%;
+  }
+
+  img {
+    position: absolute;
+    left: 0;
+    top: 0;
+    max-width: 100%;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+
+    &.default {
+      max-width: 11.6rem;
+      height: auto;
+      left: 50%;
+      top: 50%;
+      transform: translate(-50%, -50%);
+    }
+  }
+}
