+++ src/components/Product/styles.module.scss
@@ -14,108 +14,26 @@
   }
 
   @for $i from 1 through 12 {
-    .mobile-col-#{$i} {
-      width: calc(100% / #{$i} - (0.8rem * (#{$i} - 1)) / #{$i});
+    &.mobile-col-#{$i} {
+      .item {
+        width: calc(100% / #{$i} - (0.8rem * (#{$i} - 1)) / #{$i});
+      }
     }
 
     @include mixins.tabletOnly {
-      .tablet-col-#{$i} {
-        width: calc(100% / #{$i} - (1.6rem * (#{$i} - 1)) / #{$i});
+      &.tablet-col-#{$i} {
+        .item {
+          width: calc(100% / #{$i} - (1.6rem * (#{$i} - 1)) / #{$i});
+        }
       }
     }
 
     @include mixins.pcOnly {
-      .pc-col-#{$i} {
-        width: calc(100% / #{$i} - (2.4rem * (#{$i} - 1)) / #{$i});
+      &.pc-col-#{$i} {
+        .item {
+          width: calc(100% / #{$i} - (2.4rem * (#{$i} - 1)) / #{$i});
+        }
       }
     }
   }
-}
-
-.item {
-  a:hover {
-    .cover {
-      transform: translateY(-4px);
-      box-shadow: 0 10px 20px rgba(0, 0, 0, 0.04);
-    }
-  }
-  .cover {
-    overflow: hidden;
-    position: relative;
-    display: block;
-    margin-bottom: 1.6rem;
-    border-radius: 1.6rem;
-    background: var(--color-secondary-100);
-    transition: all 0.2s;
-
-    &:before {
-      content: "";
-      display: block;
-      width: 100%;
-      height: 0;
-      padding-bottom: 100%;
-    }
-
-    img {
-      position: absolute;
-      left: 0;
-      top: 0;
-      max-width: 100%;
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-
-      &.default {
-        max-width: 11.6rem;
-        height: auto;
-        left: 50%;
-        top: 50%;
-        transform: translate(-50%, -50%);
-      }
-    }
-  }
-
-  .content {
-    display: flex;
-    flex-direction: column;
-    gap: 0.6rem;
-  }
-
-  .title {
-    overflow: hidden;
-    font-size: 1.4rem;
-    font-weight: 500;
-    color: var(--color-secondary-800);
-    white-space: nowrap;
-    text-overflow: ellipsis;
-
-    em {
-      color: var(--color-primary-200);
-      font-weight: 700;
-    }
-  }
-
-  .price {
-    font-weight: 700;
-    color: var(--color-secondary-800);
-  }
-
-  .action {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-
-    .count {
-      font-size: 1.2rem;
-      color: var(--color-secondary-600);
-    }
-  }
-}
-
-.empty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 20rem;
-  color: var(--color-secondary-400);
 }
