+++ src/util/formatter.js
@@ -1,0 +1,3 @@
+export function toWon(price) {
+  return Number(price).toLocaleString() + "원";
+}
